Move ehcache into a single cache

+++ cas-server-integration-ignite/src/main/java/org/jasig/cas/ticket/registry/IgniteTicketRegistry.java
@@ -289,13 +289,13 @@ public final class IgniteTicketRegistry extends AbstractTicketRegistry {
     }
 
     @Override
-    public int sessionCount() {
+    public long sessionCount() {
         return BooleanUtils.toInteger(this.supportRegistryState, this.ticketGrantingTicketsCache
             .size(CachePeekMode.ALL), (int) super.sessionCount());
     }
 
     @Override
-    public int serviceTicketCount() {
+    public long serviceTicketCount() {
         return BooleanUtils.toInteger(this.supportRegistryState, this.serviceTicketsCache
             .size(CachePeekMode.ALL), (int) super.serviceTicketCount());
     }
